sum_test_pass

+++ average.h
@@ -1,5 +1,5 @@
 #ifndef _INCLUDE_AVERAGE_H_
 #define _INCLUDE_AVERAGE_H_
 
-int average(int*, int);
+double average(double*, int);
 #endif
