+++ sum.h
@@ -1,6 +1,5 @@
 #ifndef _INCLUDE_SUM_H
 #define _INCLUDE_SUM_H
-
-int sum(int*, int);
+double sum(double*, int);
 
 #endif
